+++ rasa_nlu/tokenizers/jieba_tokenizer.py
@@ -5,6 +5,8 @@ from __future__ import unicode_literals
 
 import glob
 import logging
+import os
+import shutil
 
 from rasa_nlu.components import Component
 from rasa_nlu.config import RasaNLUModelConfig
@@ -14,6 +16,8 @@ from typing import Any, List, Text
 
 logger = logging.getLogger(__name__)
 
+JIEBA_CUSTOM_DICTIONARY_PATH = "tokenizer_jieba"
+
 
 class JiebaTokenizer(Tokenizer, Component):
     name = "tokenizer_jieba"
@@ -21,6 +25,19 @@ class JiebaTokenizer(Tokenizer, Component):
     provides = ["tokens"]
 
     language_list = ["zh"]
+
+    defaults = {
+        "dictionary_path": None  # default don't load custom dictionary
+    }
+
+    def __init__(self, component_config=None):
+        # type: (Dict[Text, Any]) -> None
+        """Construct a new intent classifier using the MITIE framework."""
+
+        super(JiebaTokenizer, self).__init__(component_config)
+
+        # path to dictionary file or None
+        self.dictionary_path = self.component_config.get('dictionary_path')
 
     @classmethod
     def required_packages(cls):
@@ -56,6 +73,58 @@ class JiebaTokenizer(Tokenizer, Component):
     def tokenize(self, text):
         # type: (Text) -> List[Token]
         import jieba
+
+        if self.dictionary_path is not None:
+            self.load_custom_dictionary(self.dictionary_path)
+
         tokenized = jieba.tokenize(text)
         tokens = [Token(word, start) for (word, start, end) in tokenized]
         return tokens
+
+    @classmethod
+    def load(cls,
+             model_dir=None,  # type: Optional[Text]
+             model_metadata=None,  # type: Optional[Metadata]
+             cached_component=None,  # type: Optional[Component]
+             **kwargs  # type: **Any
+             ):
+        # type: (...) -> JiebaTokenizer
+
+        meta = model_metadata.for_component(cls.name)
+        relative_dictionary_path = meta.get("dictionary_path")
+
+        # get real path of dictionary path, if any
+        if relative_dictionary_path is not None:
+            dictionary_path = os.path.join(model_dir, relative_dictionary_path)
+
+            meta["dictionary_path"] = dictionary_path
+
+        return cls(meta)
+
+    @staticmethod
+    def copy_files_dir_to_dir(input_dir, output_dir):
+        # make sure target path exists
+        if not os.path.exists(output_dir):
+            os.makedirs(output_dir)
+
+        target_file_list = glob.glob("{}/*".format(input_dir))
+        for target_file in target_file_list:
+            shutil.copy2(target_file, output_dir)
+
+    def persist(self, model_dir):
+        # type: (Text) -> Optional[Dict[Text, Any]]
+        """Persist this model into the passed directory."""
+
+        model_dictionary_path = None
+
+        # copy custom dictionaries to model dir, if any
+        if self.dictionary_path is not None:
+            target_dictionary_path = os.path.join(model_dir,
+                                                  JIEBA_CUSTOM_DICTIONARY_PATH)
+            self.copy_files_dir_to_dir(self.dictionary_path,
+                                       target_dictionary_path)
+
+            # set dictionary_path of model metadata to relative path
+            model_dictionary_path = JIEBA_CUSTOM_DICTIONARY_PATH
+
+        return {"dictionary_path": model_dictionary_path}